requirements.txt
telebot==0.0.5
PyMySQL==1.0.3
python-dotenv==1.0.0